corrección y separación de archivos
