--- src/routes/historiaClinica.js
@@ -1,21 +1,0 @@
-const express = require("express");
-const {
-  getHistoriasClinicas,
-  getHistoriaClinicaById,
-  createHistoriaClinica,
-  updateHistoriaClinica,
-  deleteHistoriaClinica,
-  addNotaAHC,
-} = require("../controllers/historiaClinica.js");
-
-const router = express.Router();
-
-router.get("/", getHistoriasClinicas);
-router.get("/:id", getHistoriaClinicaById); 
-router.post("/", createHistoriaClinica);
-router.put("/:id", updateHistoriaClinica);
-router.delete("/:id", deleteHistoriaClinica);
-
-router.post("/:historiaClinicaId/notas", addNotaAHC);
-
-module.exports = router;
